requirements.txt
# LLM 기반 차트 생성 시스템 - Python 패키지 의존성
# Python 3.8+ 지원

# ================================
# 필수 패키지 (Core Dependencies)
# ================================

# 웹 프레임워크
Flask==3.0.0
Flask-Session==0.5.0
Werkzeug==3.0.1

# 데이터 처리
pandas==2.1.4
numpy==1.24.3

# 차트 생성 및 시각화
matplotlib==3.8.2

# 데이터베이스 연결
oracledb==2.0.1
SQLAlchemy==2.0.23

# HTTP 요청 및 API 통신
requests==2.31.0
urllib3==2.1.0

# 환경 설정 및 구성
python-dotenv==1.0.0

# ================================
# LLM 관련 패키지 (선택사항)
# ================================

# Hugging Face Transformers (vLLM 모드용)
transformers==4.36.2
tokenizers==0.15.0

# PyTorch (vLLM 모드용 - CPU 버전)
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Google Generative AI (Gemini API 사용 시)
google-generativeai==0.3.2
google-auth==2.23.4

# ================================
# GPU 환경 전용 (선택사항)
# ================================

# vLLM (GPU 서버 환경에서만 설치)
# vllm==0.2.6
# 주의: CUDA 11.8+ 필요, GPU 메모리 8GB+ 권장

# CUDA PyTorch (GPU 환경용 - 위의 CPU 버전과 택일)
# torch==2.1.2+cu118
# torchvision==0.16.2+cu118
# torchaudio==2.1.2+cu118
# --extra-index-url https://download.pytorch.org/whl/cu118

# ================================
# 개발 및 테스트 도구
# ================================

# 테스트 프레임워크
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# 코드 품질 및 린팅
flake8==6.1.0
black==23.11.0
isort==5.12.0

# 타입 체킹
mypy==1.7.1

# ================================
# 유틸리티 및 기타
# ================================

# 로깅 및 모니터링
coloredlogs==15.0.1

# 파일 처리
openpyxl==3.1.2
xlsxwriter==3.1.9

# 이미지 처리 (차트 생성 지원)
Pillow==10.1.0

# 날짜 및 시간 처리
python-dateutil==2.8.2
pytz==2023.3

# JSON 처리 개선
orjson==3.9.10

# ================================
# 보안 및 인증
# ================================

# 보안 강화
cryptography==41.0.7
bcrypt==4.1.2

# JWT 토큰 (향후 인증 기능용)
PyJWT==2.8.0

# ================================
# 프로덕션 환경
# ================================

# WSGI 서버 (프로덕션 배포용)
gunicorn==21.2.0
waitress==2.1.2

# 프로세스 관리
supervisor==4.2.5

# ================================
# 설치 명령어 예시
# ================================

# 기본 설치 (필수 패키지만):
# pip install Flask==3.0.0 Flask-Session==0.5.0 pandas==2.1.4 numpy==1.24.3 matplotlib==3.8.2 oracledb==2.0.1 SQLAlchemy==2.0.23 requests==2.31.0 python-dotenv==1.0.0

# 전체 설치:
# pip install -r requirements.txt

# GPU 환경 설치:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install vllm==0.2.6

# 개발 환경 설치:
# pip install -r requirements.txt
# pip install pytest pytest-flask pytest-cov flake8 black isort mypy